+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Fixed
 - Fix Performance issue on v2 Flaw endpoint with include/exclude queryparams.
+- Added lost indexes for AffectsV1 and tests to ensure they are not deleted (OSIDB-4717).
 
 ## [5.4.1] - 2026-01-15
 ### Removed